External Workshop/requirements.txt
openai==1.20.0
langchain==0.1.16
langchain-community==0.0.33
langchain-core==0.1.43
langchain-openai==0.1.3
langchain-text-splitters==0.0.1
requests==2.31.0
requests-oauthlib==2.0.0
python-dotenv==1.0.1